consolidate reltime
